pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "flatmaps"
description = "Utilities for working with cortical flat maps"
authors = [
    {name = "Connor Lane", email = "connor.lane@childmind.org"},
]
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT License"}
dependencies = [
    "numpy",
    "pycortex",
    "scikit-learn",
    "scipy",
    "shapely",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = ["pytest", "pre-commit", "ipython", "jupyter"]

[project.urls]
"Homepage" = "https://github.com/clane9/flatmaps"

[tool.setuptools_scm]
write_to = "flatmaps/_version.py"

[tool.setuptools.packages.find]
include = ["flatmaps*"]


[tool.isort]
profile = "black"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"

[tool.mypy]
no_strict_optional = true
ignore_missing_imports = true
